[fix] #181 delete a database cause a NPE in delete collection (shortcut key clash)

+++ src/main/java/org/codinjutsu/tools/mongo/view/action/DropCollectionAction.java
@@ -39,7 +39,13 @@ public class DropCollectionAction extends AnAction implements DumbAware {
 
     @Override
     public void actionPerformed(AnActionEvent anActionEvent) {
-        int result = JOptionPane.showConfirmDialog(null, String.format("Do you REALLY want to drop the '%s' collection?", mongoExplorerPanel.getSelectedCollection().getName()), "Warning", JOptionPane.YES_NO_OPTION);
+        if (mongoExplorerPanel.getSelectedCollection() == null) { //fix in case of deleting a database
+            return;
+        }
+
+        int result = JOptionPane.showConfirmDialog(null,
+                String.format("Do you REALLY want to drop the '%s' collection?",
+                        mongoExplorerPanel.getSelectedCollection().getName()), "Warning", JOptionPane.YES_NO_OPTION);
 
         if (result == JOptionPane.YES_OPTION) {
             mongoExplorerPanel.dropCollection();
